requirements.txt
playwright>=1.40.0
requests>=2.31.0
python-dotenv>=1.0.0
PyYAML>=6.0.0
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
